requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
groq>=0.9.0
python-pptx>=0.6.23
pydantic>=2.0.0
